refactor(pipeline/scrape): split ScrapeReplayAction into ScrapeReplay/ sub-folder (Phase 8.5b commit 3)
## Summary

Drains `src/Scrapers/Pipeline/Mediator/Scrape/ScrapeReplayAction.ts`
(624 raw / 324 eff LoC, §12A grandfather) into five focused sub-modules
under a new `Mediator/Scrape/ScrapeReplay/` folder, plus a barrel and a
21-LoC re-export shim at the original path.

All historical callers (4 production + 2 test) compile unchanged — the
shim preserves the exact public surface: 5 functions
(`buildMonthBody`, `generateMonthChunks`, `isMonthlyEndpoint`,
`isRangeIterable`, `replaceField`) and 2 types (`IMonthChunk`,
`JsonRecord`).

## New layout

```
Mediator/Scrape/ScrapeReplay/
  JsonTypes.ts     10 raw /   2 eff   shared JsonRecord alias
  Base64Paging.ts 107 raw /  58 eff   paging-context detect + range-iter
  JsonReplace.ts  167 raw /  83 eff   BFS field replace + base64 fallback
  RecordShape.ts  277 raw / 138 eff   account-record + month/year subst
  MonthChunking.ts 161 raw /  81 eff  monthly ISO chunk generation
  index.ts         12 raw /   6 eff   barrel
```

Old `ScrapeReplayAction.ts`: 624 → 22 raw / 8 eff (re-export shim).

## Function-cap compliance (Commit 6 lookahead)

Every new function ≤ 10 effective LoC. Counted via brace-depth state
machine matching ESLint's `max-lines-per-function` skipBlankLines +
skipComments semantics. Sub-folders are bound by §12 (max-lines: 150,
per-fn ≤ 20 currently → 10 in Commit 6); no grandfather override
applies.

## Refactor patterns (vs the original 624-LoC blob)

1. Bundled args to keep per-fn under 10 LoC despite the
   `@typescript-eslint/max-params: 3` cap and multi-line signatures:
   `IReplaceLevelArgs`, `IMonthYearArgs`, `ICompositeArgs`,
   `IReplaceLevelResult`, `IBfsReplaceCtx`.
2. `coerceToBodyType`, `findRecordKey`, `pickComposite`,
   `buildInitialState`, `applyFutureMonths`, `applyCompositeDate`,
   `applyDirectMonthYear`, `reachedEnd`, `computeChunkEndDay`,
   `findLowerHit`, `findScalarShapeHit` extracted as named helpers so
   their callers fit the per-fn budget.
3. `JsonRecord` lives in its own `JsonTypes.ts` to break what would
   otherwise be a JsonReplace ↔ Base64Paging type-import cycle.
4. Forbidden nested calls split via descriptive intermediates
   (e.g. `lowerKeys = keys.map(...)`).
5. `JsonReplace.ts` exports `replaceField` as `default` to satisfy
   `import-x/prefer-default-export` on single-export modules (existing
   project pattern — see `TxnHunt.ts`, `EndpointResolver.ts`,
   `LifoCrawl.ts`, `RawShapeTrace.ts`). The barrel re-exports it as
   named `{ default as replaceField }`.
6. Result Pattern: `findRecordKey` returns `string | false` (not
   `undefined`), per §3 `no-restricted-syntax` ban on
   `null`/`undefined` returns.
7. `applyCompositeDate`/`applyDirectMonthYear` return `true` (not
   `void`), per §3 ban on `void` return types.

## Validation

- `npx prettier --write` clean
- `npx eslint` ALL GREEN across all new files + shim
- `npx tsc --noEmit` exit 0
- `npx tsup` build OK
- `lib/index.d.ts` SHA256 BYTE-IDENTICAL to baseline
  (`221E161123230F013CB4687CE4CD51109DA5C7D579552DE5328190AD98AAB055`)
- `lib/index.cjs` 876,580 bytes (+1,161 vs Commit 2; new sub-folder
  esbuild `__esm` wrappers — same canonical pattern as Phase 8.5a)
- `npx madge --circular --extensions ts Mediator/Scrape/` 18 cycles
  (NO new cycles introduced)
- `jest --testPathPatterns="ScrapeReplay|GenericScrapeReplay"` 46/46
- `jest --testPathPatterns="Scrape"` 4850/4850 (416/416 suites)

## Test-body freeze (AC9)

Zero test files modified. `generateMonthChunks(start, end, futureMonths)`
positional signature preserved — all 14 callers (4 prod + 10 test) work
unchanged.

## Phase 8.5b ledger (commit 3 of 7)

- ✅ Commit 1: `UrlDateRange.ts` (`fce97610`)
- ✅ Commit 2: `FrozenScrapeAction.ts` (`d618168b`)
- ✅ Commit 3: `ScrapeReplayAction.ts` split (this commit)
- ⏳ Commit 4-5: `ScrapePhaseActions.ts` split
- ⏳ Commit 6: tighten §12 max-lines-per-function 20 → 10, drop §12A
- ⏳ Commit 7: re-pad canary + update `lint-guideline-coverage.ts`

Master plan: pipeline-decoupling-master-2026-05-28 / phase-8.5b
Sub-plan:    phase-8.5-canonical-10/sub-b-scrape
